Split package signature cryptography
